Deprecate email dependant methods
This commit deprecates `getLocationsForEmail` and `searchPersonByEmail`.
The commit also introduces a method `getLocationsForPerson` to be used instead.

The change is necessary due to a change in user authentification and authorization in the userDb.
The users no longer are identified by their `email` field but rather have an additional `user-id` field.